refactor: move class script.Script.Builder to script.ScriptBuilder
