+++ src/main/java/com/github/kuro46/embedscript/script/parser/Processor.java
@@ -1,8 +1,8 @@
 package com.github.kuro46.embedscript.script.parser;
 
 import com.github.kuro46.embedscript.script.ParseException;
-import com.github.kuro46.embedscript.script.Script;
 import com.github.kuro46.embedscript.script.ScriptBuffer;
+import com.github.kuro46.embedscript.script.ScriptBuilder;
 
 /**
  * Processes parse operation
@@ -12,5 +12,5 @@ public interface Processor {
 
     void setup(ScriptParser parser, ScriptBuffer source) throws ParseException;
 
-    void process(ScriptParser parser, Script.Builder builder, ScriptBuffer source) throws ParseException;
+    void process(ScriptParser parser, ScriptBuilder builder, ScriptBuffer source) throws ParseException;
 }
